fix(engine): Moved unit test too different class
